requirements.txt
azure-devops
python-gitlab==4.4.0
PyYAML